Rename ImpressionsObserver (#599)
