+++ src/main/java/io/split/android/client/service/impressions/observer/ImpressionsObserver.java
@@ -1,0 +1,10 @@
+package io.split.android.client.service.impressions.observer;
+
+import androidx.annotation.Nullable;
+
+import io.split.android.client.impressions.Impression;
+
+public interface ImpressionsObserver {
+    @Nullable
+    Long testAndSet(Impression impression);
+}
