+++ src/test/java/io/split/android/client/service/impressions/strategy/OptimizedStrategyTest.kt
@@ -4,10 +4,9 @@ import io.split.android.client.dtos.KeyImpression
 import io.split.android.client.service.executor.SplitTaskExecutionListener
 import io.split.android.client.service.executor.SplitTaskExecutor
 import io.split.android.client.service.impressions.ImpressionsCounter
-import io.split.android.client.service.impressions.ImpressionsObserver
+import io.split.android.client.service.impressions.observer.ImpressionsObserverImpl
 import io.split.android.client.service.impressions.ImpressionsRecorderTask
 import io.split.android.client.service.impressions.ImpressionsTaskFactory
-import io.split.android.client.service.sseclient.sseclient.RetryBackoffCounterTimer
 import io.split.android.client.service.synchronizer.RecorderSyncHelper
 import io.split.android.client.telemetry.model.ImpressionsDataType
 import io.split.android.client.telemetry.storage.TelemetryRuntimeProducer
@@ -29,7 +28,7 @@ class OptimizedStrategyTest {
     private lateinit var impressionsCounter: ImpressionsCounter
 
     @Mock
-    private lateinit var impressionsObserver: ImpressionsObserver
+    private lateinit var impressionsObserver: ImpressionsObserverImpl
 
     @Mock
     private lateinit var impressionsSyncHelper: RecorderSyncHelper<KeyImpression>
